+++ cmd/hmcli/paramset.go
@@ -1,0 +1,129 @@
+// SPDX-License-Identifier: MIT
+// Copyright (C) 2026 OpenCCU-Loom authors.
+
+package main
+
+import (
+	"context"
+	"errors"
+	"flag"
+	"fmt"
+	"io"
+	"net/http"
+	"sort"
+	"text/tabwriter"
+	"time"
+)
+
+type paramsetFlags struct {
+	host     string
+	token    string
+	user     string
+	password string
+	timeout  time.Duration
+	jsonOut  bool
+}
+
+func (f *paramsetFlags) bindTo(fs *flag.FlagSet) {
+	fs.StringVar(&f.host, "host", "http://localhost:8119", "daemon REST base URL")
+	fs.StringVar(&f.token, "token", "", "API bearer token")
+	fs.StringVar(&f.user, "user", "", "basic-auth username")
+	fs.StringVar(&f.password, "password", "", "basic-auth password")
+	fs.DurationVar(&f.timeout, "timeout", 60*time.Second, "request timeout")
+	fs.BoolVar(&f.jsonOut, "json", false, "emit raw JSON instead of a human-readable table")
+}
+
+func (f *paramsetFlags) client() *daemonClient {
+	return newDaemonClient(f.host, f.token, f.user, f.password, f.timeout)
+}
+
+func cmdParamset(args []string, stdout, stderr io.Writer) error {
+	if len(args) == 0 {
+		return errors.New("paramset: missing operation (try: get, set)")
+	}
+	switch args[0] {
+	case "get":
+		return cmdParamsetGet(args[1:], stdout, stderr)
+	case "set":
+		return cmdParamsetSet(args[1:], stdout, stderr)
+	default:
+		return fmt.Errorf("paramset: unknown operation %q", args[0])
+	}
+}
+
+var validParamsetKeys = map[string]bool{
+	"VALUES": true,
+	"MASTER": true,
+	"LINK":   true,
+}
+
+func cmdParamsetGet(args []string, stdout, stderr io.Writer) error {
+	fs := flag.NewFlagSet("paramset get", flag.ContinueOnError)
+	fs.SetOutput(stderr)
+	var f paramsetFlags
+	f.bindTo(fs)
+	if err := fs.Parse(args); err != nil {
+		return err
+	}
+	rest := fs.Args()
+	if len(rest) < 2 {
+		return errors.New("paramset get: usage: get <addr> <KEY>")
+	}
+	addr, key := rest[0], rest[1]
+	if !validParamsetKeys[key] {
+		return fmt.Errorf("paramset get: invalid KEY %q (must be VALUES, MASTER, or LINK)", key)
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), f.timeout)
+	defer cancel()
+
+	var params map[string]any
+	if err := f.client().getJSON(ctx, "/api/v1/devices/"+addr+"/paramsets/"+key, &params); err != nil {
+		return err
+	}
+
+	if f.jsonOut {
+		return writeJSON(stdout, params)
+	}
+
+	keys := make([]string, 0, len(params))
+	for k := range params {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+
+	tw := tabwriter.NewWriter(stdout, 0, 0, 2, ' ', 0)
+	_, _ = fmt.Fprintln(tw, "PARAM\tVALUE")
+	for _, k := range keys {
+		_, _ = fmt.Fprintf(tw, "%s\t%v\n", k, params[k])
+	}
+	return tw.Flush()
+}
+
+func cmdParamsetSet(args []string, stdout, stderr io.Writer) error {
+	fs := flag.NewFlagSet("paramset set", flag.ContinueOnError)
+	fs.SetOutput(stderr)
+	var f paramsetFlags
+	f.bindTo(fs)
+	if err := fs.Parse(args); err != nil {
+		return err
+	}
+	rest := fs.Args()
+	if len(rest) < 4 {
+		return errors.New("paramset set: usage: set <addr> <KEY> <param> <value>")
+	}
+	addr, key, param, rawVal := rest[0], rest[1], rest[2], rest[3]
+	if !validParamsetKeys[key] {
+		return fmt.Errorf("paramset set: invalid KEY %q (must be VALUES, MASTER, or LINK)", key)
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), f.timeout)
+	defer cancel()
+
+	body := map[string]any{param: coerceValue(rawVal)}
+	if err := f.client().sendJSON(ctx, http.MethodPut, "/api/v1/devices/"+addr+"/paramsets/"+key, body, nil); err != nil {
+		return err
+	}
+	_, _ = fmt.Fprintln(stdout, "ok")
+	return nil
+}